% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNet.R
\name{AcheOrdemN}
\alias{AcheOrdemN}
\title{Calcula Ordem de Operação N}
\usage{
AcheOrdemN(ReservOrdemN, ReservOrdensMenores, nID, Ord_aflu)
}
\arguments{
\item{ReservOrdemN}{Vetor de IDs de reservatórios com ordem atual.}

\item{ReservOrdensMenores}{Vetor de IDs de reservatórios com ordens inferiores.}

\item{nID}{ID do reservatório principal.}

\item{Ord_aflu}{Matriz com duas colunas: IDs dos reservatórios e seus jusantes.}
}
\value{
Vetor de IDs dos reservatórios com ordem de operação superior.
}
\description{
Determina o vetor de reservatórios com ordem de operação imediatamente superior.
}
