% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{Add_linhasAggRegHidro}
\alias{Add_linhasAggRegHidro}
\title{Adicionar Linhas de Agregação a Tabelas de Dados}
\usage{
Add_linhasAggRegHidro(
  RegHidro,
  Capacid,
  DF_list,
  L_DadosRedeBalancoMedio_AggReg,
  ind_acres
)
}
\arguments{
\item{RegHidro}{Vetor com regiões hidrográficas para cada reservatório.}

\item{Capacid}{Vetor com capacidades dos reservatórios.}

\item{DF_list}{Lista de data frames onde as linhas serão adicionadas.}

\item{L_DadosRedeBalancoMedio_AggReg}{Lista de dados agregados por região.}

\item{ind_acres}{Vetor de índices indicando onde adicionar as novas linhas.}
}
\value{
Lista de data frames com as linhas adicionadas.
}
\description{
Insere linhas de agregação de regiões hidrográficas em listas de data frames.
}
