% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{Agg_anual}
\alias{Agg_anual}
\title{Agregação Anual de Séries Temporais}
\usage{
Agg_anual(Lista_series, func_agg = mean)
}
\arguments{
\item{Lista_series}{Lista contendo séries temporais nomeadas.}

\item{func_agg}{Função de agregação a ser aplicada (padrão: `mean`).}
}
\value{
Lista com as séries agregadas anualmente. Valores para anos incompletos são definidos como `NA`.
}
\description{
Agrega séries temporais em uma base anual, aplicando uma função de agregação especificada.
}
