% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{Agg_RegHidr}
\alias{Agg_RegHidr}
\title{Agregação de Regiões Hidrográficas}
\usage{
Agg_RegHidr(RegHidro, Capacid, L_DadosRedeBalancoMedio)
}
\arguments{
\item{RegHidro}{Vetor com regiões hidrográficas para cada reservatório.}

\item{Capacid}{Vetor com capacidades dos reservatórios.}

\item{L_DadosRedeBalancoMedio}{Lista de dados médios da rede.}
}
\value{
Lista com os dados agregados por região hidrográfica.
}
\description{
Realiza a agregação de dados por região hidrográfica, considerando capacidade e dados médios.
}
