% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_OperacaoReservatorios.R
\name{Dinamica_mensal_Simplificada}
\alias{Dinamica_mensal_Simplificada}
\title{Dinâmica Mensal Simplificada de um Reservatório}
\usage{
Dinamica_mensal_Simplificada(Vol_Ini, Aflu, Evap, Retirada, Geom_Reserv)
}
\arguments{
\item{Vol_Ini}{Volume inicial no início do mês (em hm³).}

\item{Aflu}{Afluência no mês (em hm³/mês).}

\item{Evap}{Evaporação no mês (em m/mês).}

\item{Retirada}{Retirada no mês (em hm³/mês).}

\item{Geom_Reserv}{Estrutura geométrica do reservatório, incluindo curvas cota-área-volume.}
}
\value{
Vetor contendo o volume final, retirada efetiva e vertimento.
}
\description{
Calcula o volume de um reservatório ao final de um mês considerando entradas, saídas e evaporação.
}
