% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_OperacaoReservatorios.R
\name{DiscretizaGarantia}
\alias{DiscretizaGarantia}
\title{Discretiza Garantia}
\usage{
DiscretizaGarantia(nSerie, Garantia)
}
\arguments{
\item{nSerie}{Número total de séries temporais.}

\item{Garantia}{Garantia desejada (valor contínuo).}
}
\value{
Garantia ajustada ao maior valor discreto possível.
}
\description{
Ajusta o valor da garantia para o maior valor discreto possível, considerando o número de séries.
}
