% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_Download_EstacoesPlu.R
\name{Download_estacoesPlu}
\alias{Download_estacoesPlu}
\title{Dowload de dados das estações pluviometricas}
\usage{
Download_estacoesPlu(est_down, ano_i, ano_f)
}
\arguments{
\item{est_down}{Vetor de inteiros com o identificador das estações a serem baixadas}

\item{ano_i}{Ano inicial}

\item{ano_f}{Ano Final}
}
\value{
Lista com dois elementos $series e $lonlat
as séries menasais de precipitação em cada estação no período determinado e
as coordenadas geográficas das estações
}
\description{
Dowload de dados das estações pluviometricas
}
