% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{Forca_CAVs_ordcres}
\alias{Forca_CAVs_ordcres}
\title{Força as CAVs a serem todas crescentes}
\usage{
Forca_CAVs_ordcres(CAVs, ids_cor)
}
\arguments{
\item{CAVs}{Data frame contendo colunas a serem verificadas.}

\item{ids_cor}{Vetor de IDs que apresentam colunas fora de ordem.}
}
\value{
CAVs corrigidas
}
\description{
Apaga as linhas das CAVs para garantir que a CAV é crescente
}
