% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_CaracFisiogr_calc.R
\name{FUNC_area_perim}
\alias{FUNC_area_perim}
\title{Calcular Área e Perímetro de Polígonos}
\usage{
FUNC_area_perim(shpAc.UTM, idsBac)
}
\arguments{
\item{shpAc.UTM}{Lista de objetos `sf` representando os polígonos das bacias.}

\item{idsBac}{Vetor de IDs das bacias.}
}
\value{
Uma lista com vetores de área (`vAreaAc`) e perímetro (`vPerAc`).
}
\description{
Calcula a área e o perímetro de polígonos fornecidos em um objeto `sf` e retorna os resultados em km² e km, respectivamente.
}
