% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_CaracFisiogr_calc.R
\name{FUNC_arecris}
\alias{FUNC_arecris}
\title{Calcular Área de Cristalino}
\usage{
FUNC_arecris(shpAc.UTM, idsBac, shp_Cris.UTM)
}
\arguments{
\item{shpAc.UTM}{Lista de objetos `sf` representando os polígonos das bacias.}

\item{idsBac}{Vetor de IDs das bacias.}

\item{shp_Cris.UTM}{Objeto `sf` representando as áreas de cristalino.}
}
\value{
Vetor com a área de cristalino em cada bacia (em km²).
}
\description{
Calcula a área de cristalino interior às bacias fornecidas.
}
