% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_atualiz_arqNet.R
\name{FUNC_atualiz_lista_vec}
\alias{FUNC_atualiz_lista_vec}
\title{Atualiza Lista ou Vetor de Dados}
\usage{
FUNC_atualiz_lista_vec(List_tmp, List_consol)
}
\arguments{
\item{List_tmp}{Lista ou vetor. Dados dos novos reservatórios.}

\item{List_consol}{Lista ou vetor. Dados consolidados de reservatórios existentes.}
}
\value{
Lista ou vetor. Dados consolidados atualizados com os novos valores.
}
\description{
Atualiza uma lista ou vetor de dados com base em novos valores fornecidos.
}
\examples{
# Exemplo de uso
nova_lista <- FUNC_atualiz_lista_vec(List_tmp, List_consol)

}
