% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_atualiz_arqNet.R
\name{FUNC_atualiz_mat}
\alias{FUNC_atualiz_mat}
\title{Atualiza Matriz de Características}
\usage{
FUNC_atualiz_mat(mat_tmp, mat_consol, res_tmp, res_consol)
}
\arguments{
\item{mat_tmp}{Data frame. Matriz com características dos novos reservatórios.}

\item{mat_consol}{Data frame. Matriz consolidada com características de reservatórios existentes.}

\item{res_tmp}{Vetor. IDs dos novos reservatórios.}

\item{res_consol}{Vetor. IDs dos reservatórios existentes.}
}
\value{
Data frame. Matriz consolidada atualizada com os novos dados.
}
\description{
Atualiza a matriz de características de reservatórios com base em novos dados fornecidos.
}
\examples{
# Exemplo de uso
nova_matriz <- FUNC_atualiz_mat(mat_tmp, mat_consol, res_tmp, res_consol)
}
