% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_CaracFisiogr_calc.R
\name{FUNC_cad}
\alias{FUNC_cad}
\title{Calcular Capacidade de Água Disponível (CAD)}
\usage{
FUNC_cad(shpAc.UTM, idsBac, vAreaAc, shp_solos.UTM)
}
\arguments{
\item{shpAc.UTM}{Lista de objetos `sf` representando os polígonos das bacias.}

\item{idsBac}{Vetor de IDs das bacias.}

\item{vAreaAc}{Vetor de áreas das bacias (em km²).}

\item{shp_solos.UTM}{Objeto `sf` representando as áreas de solo.}
}
\value{
Vetor com a capacidade de água disponível em cada bacia.
}
\description{
Calcula a capacidade de água disponível para cada bacia com base na interseção com as áreas de solo.
}
