% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_CaracFisiogr_calc.R
\name{FUNC_CararacFisiogr_calc}
\alias{FUNC_CararacFisiogr_calc}
\title{Calcular Características Fisiográficas}
\usage{
FUNC_CararacFisiogr_calc(
  idsBac,
  arqBac,
  arqRio,
  arqCris,
  arqSolo,
  arqDecl,
  arqCN
)
}
\arguments{
\item{idsBac}{Vetor de IDs das bacias.}

\item{arqBac}{Vetor de caminhos para os arquivos das bacias.}

\item{arqRio}{Caminho para o arquivo dos rios.}

\item{arqCris}{Caminho para o arquivo das áreas de cristalino.}

\item{arqSolo}{Caminho para o arquivo das áreas de solo.}

\item{arqDecl}{Caminho para o arquivo raster de declividade.}

\item{arqCN}{Caminho para o arquivo raster CN.}
}
\value{
Matriz com várias características fisiográficas das bacias.
}
\description{
Calcula várias características fisiográficas das bacias de acúmulo de água.
}
