% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_CaracFisiogr_calc.R
\name{FUNC_compdren}
\alias{FUNC_compdren}
\title{Calcular Comprimento Total de Drenagem}
\usage{
FUNC_compdren(shpAc.UTM, idsBac, shp_Rio.UTM)
}
\arguments{
\item{shpAc.UTM}{Lista de objetos `sf` representando os polígonos das bacias.}

\item{idsBac}{Vetor de IDs das bacias.}

\item{shp_Rio.UTM}{Objeto `sf` representando os rios.}
}
\value{
Vetor com o comprimento total dos rios em cada bacia (em km).
}
\description{
Calcula o comprimento total dos rios interiores às bacias fornecidas.
}
