% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_regionalizParam_aplica.R
\name{FUNC_estproxKNN}
\alias{FUNC_estproxKNN}
\title{Encontra as estações mais próximas de acordo com características selecionadas}
\usage{
FUNC_estproxKNN(CaracEst, CaracAcu, carac_compara, pesos_carac, n_viz = 5)
}
\arguments{
\item{CaracEst}{Data.frame com as características das estações.}

\item{CaracAcu}{Data.frame com as características acumuladas.}

\item{carac_compara}{Vetor com os nomes das características a serem usadas na comparação.}

\item{pesos_carac}{Vetor de pesos para ponderação das características.}

\item{n_viz}{Número de vizinhos (default = 5)}
}
\value{
Lista com as estações mais próximas para cada entrada em CaracAcu.
}
\description{
Esta função utiliza o método K-Nearest Neighbors (KNN) para identificar as cinco
estações mais próximas em termos de características selecionadas e ponderadas.
}
\examples{
CaracEst <- data.frame(A = rnorm(10), B = rnorm(10), C = rnorm(10))
CaracAcu <- data.frame(A = rnorm(3), B = rnorm(3), C = rnorm(3))
carac_compara <- c("A", "B", "C")
pesos_carac <- c(A = 1, B = 1, C = 1)
FUNC_estproxKNN(CaracEst, CaracAcu, carac_compara, pesos_carac)
}
