% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_thiessen_internos_IDW.R
\name{FUNC_ETPmed_thiessen}
\alias{FUNC_ETPmed_thiessen}
\title{Calcula a ETP média ponderada por Thiessen para várias bacias}
\usage{
FUNC_ETPmed_thiessen(MatETP, pontos, arqBac, idsBac)
}
\arguments{
\item{MatETP}{Matriz com valores de ETP média mensal por ponto. Nomeadas nas linhas pelo id dos ponstos.}

\item{pontos}{Data.frame com as coordenadas $x e $y e identificador do ponto/posto $id.}

\item{arqBac}{Lista de caminhos para os arquivos das shapes das bacias hidrográficas.}

\item{idsBac}{Vetor de IDs das bacias hidrográficas.}
}
\value{
Lista onde cada elemento corresponde à ETP média ponderada para uma bacia.
}
\description{
Esta função calcula a evapotranspiração potencial média (ETP) para
várias bacias hidrográficas, utilizando a ponderação pelos pesos de Thiessen
calculados para cada bacia.
}
