% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_MediasRegionais_SerSinteticas.R
\name{FUNC_fragmentos}
\alias{FUNC_fragmentos}
\title{Calcular Fragmentos Mensais de Séries de Vazão}
\usage{
FUNC_fragmentos(LVaz)
}
\arguments{
\item{LVaz}{Lista de séries temporais de vazão (`numeric`), onde cada elemento representa uma série de dados mensais de vazão.}
}
\value{
Uma matriz onde cada linha representa uma série de vazão e cada coluna, o fragmento médio para cada mês (valores normalizados em relação à vazão anual média).
}
\description{
Essa função calcula os fragmentos mensais de séries de vazão, que são as proporções médias da vazão mensal em relação à vazão anual média.
}
