% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_thiessen_internos_IDW.R
\name{FUNC_IDW}
\alias{FUNC_IDW}
\title{Interpolação pelo ponderada pelo inverso da distância (IDW)}
\usage{
FUNC_IDW(series_est, coord_est, grid)
}
\arguments{
\item{listPrec}{Lista onde cada elemento é um vetor com os valores de precipitação para cada ponto.}

\item{pontos}{Data.frame com as coordenadas $x e $y e identificador do ponto/posto $id.}

\item{arqBac}{Lista de caminhos para os arquivos das shapes das bacias hidrográficas.}

\item{idsBac}{Vetor de IDs das bacias hidrográficas.}
}
\value{
Lista onde cada elemento corresponde à precipitação média para uma bacia,
        considerando apenas os pontos internos.
}
\description{
Interpolação pelo ponderada pelo inverso da distância (IDW)
}
