% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_atualiz_arqNet.R
\name{FUNC_ler_CAVs_banco}
\alias{FUNC_ler_CAVs_banco}
\title{Leitura de CAVs a Partir do Banco de Dados}
\usage{
FUNC_ler_CAVs_banco(
  id_res,
  dbname = "func",
  host = "flora.funceme.br",
  port = 5432,
  user = "consulta"
)
}
\arguments{
\item{id_res}{Vetor. IDs dos reservatórios cujos dados de CAVs serão lidos.}

\item{dbname}{Nome do banco de dados. Padrão: "func".}

\item{host}{Host do banco de dados. Padrão: "flora.funceme.br".}

\item{port}{Porta do banco de dados. Padrão: 5432.}

\item{user}{Usuário para conexão. Padrão: "consulta".}
}
\value{
Data frame. Matriz com os dados de CAVs, incluindo colunas de ID, cota, volume (em m³) e área (em km²).
}
\description{
Realiza a leitura dos dados de CAVs para um conjunto de reservatórios a partir de um banco de dados PostgreSQL.
}
\examples{
# Exemplo de uso
cavs <- FUNC_ler_CAVs_banco(c(1, 2, 3))

}
