% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_atualiz_arqNet.R
\name{FUNC_ler_cotaver}
\alias{FUNC_ler_cotaver}
\title{Leitura de cota do vertedouro a Partir do Banco de Dados}
\usage{
FUNC_ler_cotaver(
  id_res,
  dbname = "func",
  host = "flora.funceme.br",
  port = 5432,
  user = "consulta"
)
}
\arguments{
\item{id_res}{Vetor. IDs dos reservatórios cujos dados de cota do vertedouro serão lidos.}

\item{dbname}{Nome do banco de dados. Padrão: "func".}

\item{host}{Host do banco de dados. Padrão: "flora.funceme.br".}

\item{port}{Porta do banco de dados. Padrão: 5432.}

\item{user}{Usuário para conexão. Padrão: "consulta".}
}
\value{
Data frame. Matriz incluindo colunas de ID e cota do vertedouro.
}
\description{
Realiza a leitura dos dados de cota do vertedouro para um conjunto de reservatórios a partir de um banco de dados PostgreSQL.
}
