% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_CaracFisiogr_calc.R
\name{FUNC_meanrst}
\alias{FUNC_meanrst}
\title{Calcular Média de Raster}
\usage{
FUNC_meanrst(shpAc.coordGeo, idsBac, rst)
}
\arguments{
\item{shpAc.coordGeo}{Lista de objetos `sf` representando os polígonos das bacias.}

\item{idsBac}{Vetor de IDs das bacias.}

\item{rst}{Objeto `raster` a ser analisado.}
}
\value{
Vetor com a média dos valores do raster em cada bacia.
}
\description{
Calcula a média dos valores de um raster que intersectam as bacias fornecidas.
}
