% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_regionalizParam_aplica.R
\name{FUNC_paramKNN}
\alias{FUNC_paramKNN}
\title{Obtém Parâmetros KNN}
\usage{
FUNC_paramKNN(est_prox, ParamEstFlu)
}
\arguments{
\item{est_prox}{Vetor com os índices ou identificadores das estações mais próximas.}

\item{ParamEstFlu}{Data frame contendo os parâmetros das estações fluviométricas.}
}
\value{
Data frame com os parâmetros das estações selecionadas.
}
\description{
Seleciona os parâmetros das estações mais próximas para o método KNN.
}
