% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_regionalizParam_aplica.R
\name{FUNC_paramML}
\alias{FUNC_paramML}
\title{Calcula Parâmetros por Modelo Linear}
\usage{
FUNC_paramML(Carac, Prec_med)
}
\arguments{
\item{Carac}{Data frame com características das estações (e.g., "Cristalino", "CAD_mm").}

\item{Prec_med}{Vetor com a precipitação média anual das estações.}
}
\value{
Data frame com os parâmetros calculados: SAT, PES, CREC e K.
}
\description{
Calcula os parâmetros para modelagem hidrológica usando características do solo e dados de precipitação.
}
