% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_thiessen_internos_IDW.R
\name{FUNC_Precmed_internos}
\alias{FUNC_Precmed_internos}
\title{Calcula a precipitação média para várias bacias hidrográficas utilizando apenas pontos internos}
\usage{
FUNC_Precmed_internos(listPrec, pontos, arqBac, idsBac)
}
\arguments{
\item{listPrec}{Lista onde cada elemento é um vetor com os valores de precipitação para cada ponto.}

\item{pontos}{Data.frame com as coordenadas $x e $y e identificador do ponto/posto $id.}

\item{arqBac}{Lista de caminhos para os arquivos das shapes das bacias hidrográficas.}

\item{idsBac}{Vetor de IDs das bacias hidrográficas.}
}
\value{
Lista onde cada elemento corresponde à precipitação média para uma bacia,
        considerando apenas os pontos internos.
        @export
}
\description{
Esta função calcula a precipitação média em cada bacia hidrográfica, considerando apenas
os pontos que estão localizados dentro de cada bacia. A precipitação média é calculada
para cada bacia utilizando a lista de precipitações fornecida.
}
