% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_thiessen_internos_IDW.R
\name{FUNC_pts_internos}
\alias{FUNC_pts_internos}
\title{Identifica pontos que estão dentro de uma shape}
\usage{
FUNC_pts_internos(pontos, shp_aq)
}
\arguments{
\item{pontos}{Data.frame com as coordenadas $x e $y e identificador do ponto/posto $id.}

\item{shp_aq}{Spatial Feature (lida com `read_sf`) ou caminho para o arquivo com a shape.}
}
\value{
Vetor de IDs dos pontos que estão dentro da shape.
}
\description{
Esta função verifica quais pontos de um conjunto de coordenadas estão localizados
dentro de uma área delimitada por uma shape (polígono).
}
\examples{
pontos <- data.frame(x = c(1, 2), y = c(3, 4), id = c("A", "B"))
shp <- st_read("path_to_shapefile.shp")
FUNC_pts_internos(pontos, shp)
}
