% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNet.R
\name{FUNC_resFimRede_selec}
\alias{FUNC_resFimRede_selec}
\title{Seleciona Reservatórios Finais Modificados}
\usage{
FUNC_resFimRede_selec(Ord_aflu, res_mod)
}
\arguments{
\item{Ord_aflu}{Matriz com duas colunas: IDs dos reservatórios e seus jusantes.}

\item{res_mod}{Vetor com IDs dos reservatórios modificados.}
}
\value{
Vetor de IDs dos reservatórios finais selecionados.
}
\description{
Identifica reservatórios finais de rede que fazem parte de cascatas com reservatórios modificados.
}
