% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNet.R
\name{FUNC_resRede_selec}
\alias{FUNC_resRede_selec}
\title{Seleciona Reservatórios de Rede}
\usage{
FUNC_resRede_selec(Ord_aflu, resFimRede_selec)
}
\arguments{
\item{Ord_aflu}{Matriz com duas colunas: IDs dos reservatórios e seus jusantes.}

\item{resFimRede_selec}{Vetor com IDs dos reservatórios finais selecionados.}
}
\value{
Vetor de IDs dos reservatórios que fazem parte das cascatas selecionadas.
}
\description{
Retorna os IDs de todos os reservatórios que pertencem a cascatas selecionadas.
}
