% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_MediasRegionais_SerSinteticas.R
\name{FUNC_SerSintecia}
\alias{FUNC_SerSintecia}
\title{Gerar Série Sintética de Vazão}
\usage{
FUNC_SerSintecia(vecFrag, Precmed, Area, CE, CV, dat_ini, dat_fin)
}
\arguments{
\item{vecFrag}{Vetor de fragmentos mensais, representando as proporções médias da vazão mensal em relação à vazão anual média.}

\item{Precmed}{Precipitação média anual (em mm).}

\item{Area}{Área da bacia hidrográfica (em km²).}

\item{CE}{Coeficiente de escoamento (adimensional).}

\item{CV}{Coeficiente de variação da vazão anual (adimensional).}

\item{dat_ini}{Data de início da série (formato "yyyy-mm-dd").}

\item{dat_fin}{Data de término da série (formato "yyyy-mm-dd").}
}
\value{
Um vetor numérico com a série sintética de vazão mensal, onde os nomes do vetor são as datas correspondentes.
}
\description{
Essa função gera uma série sintética de vazão mensal com base em parâmetros fornecidos e fragmentos mensais.
}
