% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_thiessen_internos_IDW.R
\name{FUNC_thiessen}
\alias{FUNC_thiessen}
\title{Calcula os pesos de Thiessen para pontos dentro de uma shape}
\usage{
FUNC_thiessen(pontos, shp_aq)
}
\arguments{
\item{pontos}{Data.frame com as coordenadas $x e $y e identificador do ponto/posto $id.}

\item{shp_aq}{Spatial Feature (lida com `read_sf`) ou caminho para o arquivo com a shape.}
}
\value{
Vetor de pesos, onde cada peso corresponde à fração da área de Thiessen
        associada a cada ponto.
}
\description{
Esta função calcula os pesos de Thiessen para um conjunto de pontos dentro de uma
área delimitada por uma shape (polígono). Os pesos são usados para determinar
a contribuição relativa de cada ponto em relação à área total.
}
\examples{
pontos <- data.frame(x = c(1, 2), y = c(3, 4), id = c("A", "B"))
shp <- st_read("path_to_shapefile.shp")
FUNC_thiessen(pontos, shp)
}
