% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_OperacaoReservatorios.R
\name{FuncGarantia}
\alias{FuncGarantia}
\title{Calcula Garantia do Reservatório}
\usage{
FuncGarantia(Acude, Ret, SerieAfluEvap, Voliniperc = 0.5, garantia_anual = F)
}
\arguments{
\item{Acude}{Lista contendo informações do reservatório, incluindo geometria (Vol_Min, Vol_Max_Operacional).}

\item{Ret}{Retirada mensal em metros cúbicos por segundo (m³/s).}

\item{SerieAfluEvap}{Data frame com séries de afluências (m³/s) e evaporações (mm/mês).}

\item{Voliniperc}{Percentual do volume útil inicial (padrão: 0.5).}
}
\value{
Garantia do reservatório (valor entre 0 e 1).
}
\description{
Calcula a garantia de atendimento para um reservatório, dado um valor de retirada.
}
