% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{Get_AfluTot_fromListSimul}
\alias{Get_AfluTot_fromListSimul}
\title{Extrair Dados de Afluência Total de Listas de Simulação}
\usage{
Get_AfluTot_fromListSimul(ListSimul)
}
\arguments{
\item{ListSimul}{Lista contendo resultados de simulações.}
}
\value{
Lista de séries temporais de afluência total para cada simulação.
}
\description{
Extrai dados de afluência total a partir de uma lista de simulações.
}
