% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_OperacaoReservatorios.R
\name{Interpola}
\alias{Interpola}
\title{Interpolação Linear Simples}
\usage{
Interpola(vecIn, vecOut, input)
}
\arguments{
\item{vecIn}{Vetor de valores de entrada (ex.: cotas).}

\item{vecOut}{Vetor de valores de saída (ex.: volumes ou áreas).}

\item{input}{Valor de entrada para o qual será calculada a interpolação.}
}
\value{
Valor interpolado correspondente.
}
\description{
Essa função realiza uma interpolação linear para determinar o valor correspondente a uma entrada.
}
