% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_OperacaoReservatorios.R
\name{LerAcude}
\alias{LerAcude}
\title{Ler Dados de um Açude}
\usage{
LerAcude(nID, DadosRes, DadosCAV)
}
\arguments{
\item{nID}{ID do reservatório a ser buscado.}

\item{DadosRes}{Data frame contendo informações gerais dos reservatórios (como cota vertente e nome).}

\item{DadosCAV}{Data frame contendo informações da relação cota-área-volume dos reservatórios.}
}
\value{
Uma lista contendo a identificação do reservatório e suas características geométricas.
}
\description{
Essa função extrai os dados geométricos e de identificação de um reservatório (açude) com base no ID fornecido.
}
