% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{MedMensalEst_para_SerieRes}
\alias{MedMensalEst_para_SerieRes}
\title{Converter Matrizes de Médias Mensais para Séries Temporais}
\usage{
MedMensalEst_para_SerieRes(Mat_MedMensal, ResEst, Datas)
}
\arguments{
\item{Mat_MedMensal}{Data frame com médias mensais, onde colunas de 3 a 14 representam meses.}

\item{ResEst}{Data frame contendo informações sobre estações (códigos e IDs).}

\item{Datas}{Vetor de datas associado às séries.}
}
\value{
Lista de séries temporais para cada estação.
}
\description{
Converte uma matriz de médias mensais em uma lista de séries temporais associadas a estações.
}
