% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNet.R
\name{MontaCascata}
\alias{MontaCascata}
\title{Monta Cascata de Reservatórios}
\usage{
MontaCascata(nID, Ord_aflu)
}
\arguments{
\item{nID}{ID do reservatório principal.}

\item{Ord_aflu}{Matriz com duas colunas: IDs dos reservatórios e seus jusantes.}
}
\value{
Matriz com os IDs dos reservatórios e suas respectivas ordens de operação.
}
\description{
Gera a lista de reservatórios a montante e suas ordens de operação.
}
