% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNet.R
\name{Montantes}
\alias{Montantes}
\title{Identifica Montantes Imediatos}
\usage{
Montantes(nID, Ord_aflu)
}
\arguments{
\item{nID}{ID do reservatório principal.}

\item{Ord_aflu}{Matriz com duas colunas: IDs dos reservatórios e seus jusantes.}
}
\value{
Vetor de IDs dos reservatórios imediatamente a montante.
}
\description{
Retorna os reservatórios imediatamente a montante de um dado reservatório.
}
