% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNetworkSimulation.R
\name{MultiplReservoirNetworkSimulation}
\alias{MultiplReservoirNetworkSimulation}
\title{Simulação de Múltiplas Redes de Reservatórios}
\usage{
MultiplReservoirNetworkSimulation(
  vec_ID_final,
  vecGarantia = c(0.9, 0.95, 0.98),
  AfluInc,
  EvapSerie,
  Ord_aflu,
  DadosRes,
  DadosCAV,
  vecVoliniperc,
  garantia_anual = FALSE
)
}
\arguments{
\item{vec_ID_final}{Vetor com os IDs dos reservatórios finais na rede.}

\item{vecGarantia}{Vetor com os níveis de garantia a serem simulados (valores entre 0 e 1). Padrão: `c(0.9, 0.95, 0.98)`.}

\item{AfluInc}{Matriz ou lista representando as séries de incrementos de afluência para os reservatórios.}

\item{EvapSerie}{Série temporal de taxas de evaporação (em mm/mês).}

\item{Ord_aflu}{Matriz com duas colunas: IDs dos reservatórios e seus jusantes.}

\item{DadosRes}{Lista ou estrutura contendo informações dos reservatórios (ex.: volumes máximos, mínimos, operacionais).}

\item{DadosCAV}{Dados relacionados às curvas de área-volume para os reservatórios.}

\item{vecVoliniperc}{Vetor com os percentuais de volume inicial para os reservatórios.}
}
\value{
Lista de listas de simulações (`LLSimul`). Cada sublista corresponde a um nível de garantia e contém os resultados das simulações para cada reservatório da rede.
}
\description{
Realiza simulações em redes de reservatórios para diferentes níveis de garantia especificados.
}
\details{
Para cada nível de garantia especificado em `vecGarantia`, a função realiza as seguintes operações:
- Itera pelos reservatórios finais fornecidos em `vec_ID_final`.
- Chama a função `ReservoirNetworkSimulation` para realizar a simulação da rede associada ao reservatório final com o nível de garantia especificado.
- Consolida os resultados em uma lista organizada por IDs dos reservatórios e níveis de garantia.
}
\examples{
\dontrun{
vec_ID_final <- c(1, 2, 3)
vecGarantia <- c(0.9, 0.95)
AfluInc <- matrix(runif(100), nrow=10)
EvapSerie <- runif(10, 5, 15)
Ord_aflu <- matrix(c(1, 0, 2, 1, 3, 1), ncol=2, byrow=TRUE)
DadosRes <- list(Vol_Max = c(100, 200, 300), Vol_Min = c(10, 20, 30))
DadosCAV <- list()
vecVoliniperc <- c(0.5, 0.7, 0.6)
result <- MultiplReservoirNetworkSimulation(vec_ID_final, vecGarantia,
                                            AfluInc, EvapSerie, Ord_aflu,
                                            DadosRes, DadosCAV, vecVoliniperc)
}
}
\seealso{
`ReservoirNetworkSimulation`
}
