% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{Org_serie_tab}
\alias{Org_serie_tab}
\title{Organizar Série Temporal em Tabela Mensal}
\usage{
Org_serie_tab(serie, datas)
}
\arguments{
\item{serie}{Vetor da série temporal.}

\item{datas}{Vetor de datas associado à série.}
}
\value{
Data frame com colunas representando meses e linhas representando anos.
}
\description{
Reorganiza uma série temporal em formato de tabela com colunas representando meses.
}
