% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNetworkSimulation.R
\name{QGarantia_withoutNet}
\alias{QGarantia_withoutNet}
\title{Title}
\usage{
QGarantia_withoutNet(
  AfluInc,
  EvapSerie,
  Ord_aflu,
  ID,
  vecVoliniperc,
  Garantia = 0.9,
  garantia_anual = FALSE
)
}
\arguments{
\item{AfluInc}{Lista nomeada(indice doas açudes) com a série de afluÇencias incrementais
(nomeadas com as datas). Todas as séries com o memso comprimento e correpontende às séries de Evap}

\item{EvapSerie}{Lista nomeada (indice doas açudes na mesma ordem que em AfluInc) com a série de
evaporação dos lagos (nomeadas com as datas). Todas as séries com o memso comprimento e correpontende
às séries de Evap}

\item{Ord_aflu}{Matriz com a primeira coluna referente aos índices dos açudes e com a segunda coluna
igual ao indice do reservatório imediatamente à jusante. Caso não haja reservatório a jusante o valor
deve ser 0.}

\item{ID}{}
}
\value{
Vazão regularizada
}
\description{
Title
}
