% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNetworkSimulation.R
\name{ReservoirNetworkSimulation}
\alias{ReservoirNetworkSimulation}
\title{Title}
\usage{
ReservoirNetworkSimulation(
  ID_final,
  Garantia = 0.9,
  AfluInc,
  EvapSerie,
  Ord_aflu,
  DadosRes,
  DadosCAV,
  vecVoliniperc,
  returnSim = T,
  garantia_anual = FALSE
)
}
\arguments{
\item{ID_final}{Indice do reservatório final da cascata a ser simulada}

\item{AfluInc}{Lista nomeada(indice doas açudes) com a série de afluÇencias incrementais
(nomeadas com as datas). Todas as séries com o memso comprimento e correpontende às séries de Evap}

\item{EvapSerie}{Lista nomeada (indice doas açudes na mesma ordem que em AfluInc) com a série de
evaporação dos lagos (nomeadas com as datas). Todas as séries com o memso comprimento e correpontende
às séries de Evap}

\item{Ord_aflu}{Matriz com a primeira coluna referente aos índices dos açudes e com a segunda coluna
igual ao indice do reservatório imediatamente à jusante. Caso não haja reservatório a jusante o valor
deve ser 0.}

\item{vecVoliniperc}{Vetor com percentual da capacidade no inicio da simulação de cada reservatório.
Nomeado com o ID do reservatório na mesma ordem que AfluInc e EvapSerie.}
}
\value{
Resultados da simulação
}
\description{
Title
}
