% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_ReservoirNet.R
\name{Simpl.Jusantes}
\alias{Simpl.Jusantes}
\title{Simplifica Lista de Jusantes}
\usage{
Simpl.Jusantes(Ord_aflu, IDres)
}
\arguments{
\item{Ord_aflu}{Matriz com duas colunas: a primeira contendo os IDs dos reservatórios e a segunda os IDs dos jusantes.}

\item{IDres}{Vetor de IDs dos reservatórios predefinidos.}
}
\value{
Matriz simplificada com jusantes considerando apenas os reservatórios predefinidos.
}
\description{
Reorganiza a lista de jusantes considerando apenas os reservatórios predefinidos (com dados disponíveis).
}
