% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_OperacaoReservatorios.R
\name{SimulaAcudeMensal}
\alias{SimulaAcudeMensal}
\title{Simula Reservatório Mensalmente}
\usage{
SimulaAcudeMensal(Acude, SerieAfluEvapRet, Vol_ini_perc)
}
\arguments{
\item{Acude}{Lista contendo informações do reservatório.}

\item{SerieAfluEvapRet}{Data frame com séries de afluências (m³/s), evaporações (mm/mês) e retiradas (m³/s).}

\item{Vol_ini_perc}{Percentual do volume útil inicial.}
}
\value{
Data frame com resultados da simulação, incluindo volumes, retiradas e vertimentos.
}
\description{
Realiza a simulação do comportamento de um reservatório ao longo de uma série temporal mensal.
}
