% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_atualiz_CAVs_redeConsol}
\alias{SisMEVAZ_atualiz_CAVs_redeConsol}
\title{Atualização da rede consolidada - nova geometria dos reservatórios}
\usage{
SisMEVAZ_atualiz_CAVs_redeConsol(
  diretorio_de_dados = paste0(getwd(), "/../../dados"),
  arq_CAVs = NULL,
  arq_cotaVert = NULL
)
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote (descompactado)}

\item{arq_CAVs}{Caractere com o caminho para o arquivo com novas CAVs. Se NULL busca CAVs no banco de dados.}

\item{arq_cotaVert}{Caractere com o caminho para o arquivo com novas cotas do vertedouro.Se NULL busca CAVs no banco de dados.}
}
\value{
"Atualização de CAVs finalizada"
}
\description{
Simulação da MEVAZ com novas CAVs e capacidades (a partir da cota do vertedouro)
dos reservatórios e atualiza a rede consolidada
}
\details{
Salva a versão anterior da rede consolidada no diretório
"diretorio_de_dados/RededeReservatorios_consolidada/versoes_anteriores" e atualiza o log
}
