% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_atualiz_Pluviometria_redeConsol}
\alias{SisMEVAZ_atualiz_Pluviometria_redeConsol}
\title{Atualização temporal das séries}
\usage{
SisMEVAZ_atualiz_Pluviometria_redeConsol(
  diretorio_de_dados = paste0(getwd(), "/../../dados")
)
}
\value{
"Atualização temporal das séries finalizada"
}
\description{
Simulação da MEVAZ com a atilização temporal das séries, a partir da atualização da precipitação
 até o ano anterior ao corrente
}
