% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_atualiz_redeTeste_para_redeConsol}
\alias{SisMEVAZ_atualiz_redeTeste_para_redeConsol}
\title{Atualização da rede consolidada - inclusão de reservatório}
\usage{
SisMEVAZ_atualiz_redeTeste_para_redeConsol(
  diretorio_de_dados = paste0(getwd(), "/../../dados")
)
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote}
}
\value{
"Simlação da rede teste finalizada"
}
\description{
Substitui a rede teste como nova rede consolidada, substituindo os arquivos no
diretório "diretorio_de_dados/RededeReservatorios_consolidada"
}
\details{
Salva a veãrso anterior da rede consolidada no diretório
"diretorio_de_dados/RededeReservatorios_consolidada/versoes_anteriores" e atualiza o log
}
\seealso{
\code{\link{FUNC_redeTeste}}
}
