% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_comparaVersoes}
\alias{SisMEVAZ_comparaVersoes}
\title{Compara duas versões}
\usage{
SisMEVAZ_comparaVersoes(
  diretorio_de_dados = paste0(getwd(), "/../../dados"),
  versao_ref = "redeConsol",
  versao_nov = "redeTeste",
  savexlsx = F
)
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote}

\item{versao_ref}{Caractere com a versão de referência ("redeConsol", "redeTeste", "Versao_i" ou "Teste_i" com i inteiro>=0)}

\item{versao_nov}{Caractere com a versão a ser analisada ("redeConsol", "redeTeste", "Versao_i" ou "Teste_i" com i inteiro>=0)}
}
\value{
arquivo de comparação formatado
}
\description{
Calcula as mudanças de variáveis-chave da MEVAZ
(precipitação, capacidade, afluência, Q90, Q95 e Q98) entre duas versões
}
