% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ_dependencias_exterCRAN.R
\name{SisMEVAZ_instala_dependencias_externas}
\alias{SisMEVAZ_instala_dependencias_externas}
\title{Instala dependências externas ao CRAN}
\usage{
SisMEVAZ_instala_dependencias_externas()
}
\description{
Instala bibliotecas hydrobr e phylin, externas ao CRAN
}
