% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_outros_testes}
\alias{SisMEVAZ_outros_testes}
\title{Outros testes}
\usage{
SisMEVAZ_outros_testes(
  diretorio_de_dados = paste0(getwd(), "/../../dados"),
  versao_ref = "redeConsol",
  ano_ini = 1911,
  ano_fin = year(Sys.time()) - 1,
  vol_ini = 0.5,
  garantia_anual = F
)
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote}

\item{versao_ref}{Caractere com a versão de referência ("redeConsol", "redeTeste", "Versao_i" ou "Teste_i" com i inteiro>=0)}
}
\value{
"Restauração finalizada"
}
\description{
Realiza outros testes sobre uma versão específica, tais como: diferentes períodos de simulação da cascata,
diferentes inicializações dos reservatórios, otimização e simulações com garantias anuais
}
\details{
Salva a veãrso anterior da rede consolidada no diretório
"diretorio_de_dados/RededeReservatorios_consolidada/versoes_anteriores" e atualiza o log
}
