% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_print_Series}
\alias{SisMEVAZ_print_Series}
\title{Escrever arquivos de séries da simulação}
\usage{
SisMEVAZ_print_Series(
  diretorio_de_dados = paste0(getwd(), "/../../dados"),
  versao_print = "redeTeste"
)
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote}

\item{versao_ref}{Caractere com a versão de referência ("redeConsol", "redeTeste", "Versao_i" ou "Teste_i" com i inteiro>=0)}
}
\value{
"Arquivos de séries de afluência escritos"
}
\description{
Função escreve arquivos .xlsx com séries de vazão afluente
Pode ser aplicada para a rede de reservatórios em teste ou para a
rede de reservatórios consolidada (versão atual ou anteriores).
}
\seealso{
\code{\link{FUNC_print_Sintese}}
}
