% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_print_Sintese}
\alias{SisMEVAZ_print_Sintese}
\title{Escrever arquivos com séries de vazão}
\usage{
SisMEVAZ_print_Sintese(
  diretorio_de_dados = paste0(getwd(), "/../../dados"),
  versao_print = "redeTeste"
)
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote}

\item{versao_print}{Caractere com a versão de referência ("redeConsol", "redeTeste", "Versao_i" ou "Teste_i" com i inteiro>=0)}
}
\value{
"Arquivo Síntese.xlsx escrito"
}
\description{
Função escreve arquivos .xlsx com séries de vazão afluente (incremental e total)
para cada reservatório, a partir dos resultados simulados.
Pode ser aplicada para a rede de reservatórios em teste ou para a
rede de reservatórios consolidada.
}
\seealso{
\code{\link{SisMEVAZ_print_Series}}
}
