% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_redeTeste}
\alias{SisMEVAZ_redeTeste}
\title{Simulação de Rede de reservatórios em teste}
\usage{
SisMEVAZ_redeTeste(diretorio_de_dados = paste0(getwd(), "/../../dados"))
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote}
}
\value{
Os resultados da inclusão do(s) novo(s) reservatório(s) são salvos no diretório
"diretorio_de_dados/RededeReservatorios_em_teste/output". Para gerar os arquivos xlsx de
saída veja SisMEVAZ_print_Sintese e SisMEVAZ_print_Series
}
\description{
Permite replicar a MEVAZ (Metodologia de Cálculo de Vazões)
incuindo um ou mais reservatórios à rede original.
Os dados dos novos reservatórios e daqueles reservatórios cujos dados forem modificados
por essa inclusão (bacia incremental ou ordem de afluência) devem ser inseridos no
diretório "diretorio_de_dados/RededeReservatorios_em_teste/input "
}
\seealso{
\code{\link{SisMEVAZ_print_Sintese}}

\code{\link{SisMEVAZ_print_Series}}
}
