% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sis-MEVAZ.R
\name{SisMEVAZ_restaura_versao}
\alias{SisMEVAZ_restaura_versao}
\title{Restaura versão da rede consolidada}
\usage{
SisMEVAZ_restaura_versao(
  diretorio_de_dados = paste0(getwd(), "/../../dados"),
  versao
)
}
\arguments{
\item{diretorio_de_dados}{Caractere com o caminho do pacote}

\item{versao}{Inteiro maior ou igual a 1 indicando a versao a ser restaurada}
}
\value{
"Restauração finalizada"
}
\description{
Restaura versão anterior da rede consolidada, substituindo os arquivos no
diretório "diretorio_de_dados/RededeReservatorios_consolidada"
}
\details{
Salva a veãrso anterior da rede consolidada no diretório
"diretorio_de_dados/RededeReservatorios_consolidada/versoes_anteriores" e atualiza o log
}
