% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_SMAPmes.R
\name{SMAPMes.inic}
\alias{SMAPMes.inic}
\title{Funcao para criar objeto com os valores iniciais para rodar o SMAP}
\usage{
SMAPMes.inic(Area, EbInic, TuInic)
}
\arguments{
\item{Area}{Area da bacia (km2)}

\item{EbInic}{Vazao Basica Inicial em m3/s (0.95 da vazao minima), default = 0}

\item{TuInic}{Teor de Umidade inicial, default = 0.3}
}
\value{
List contendo esses parametros
}
\description{
Funcao para criar objeto com os valores iniciais para rodar o SMAP
}
