% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_SMAPmes.R
\name{SMAPMes.param}
\alias{SMAPMes.param}
\title{Funcao para criar um objeto com os parametros constantes para rodar o SMAP}
\usage{
SMAPMes.param(SAT, PES, CREC, K)
}
\arguments{
\item{SAT}{Capacidade maxima do solo (400 < SAT <5000)}

\item{PES}{Parametro de Escoamento (0.1 < PES < 10)}

\item{CREC}{Parametro de Recarga (0 < CREC < 70)}

\item{K}{Constante (1 < K < 6)}
}
\value{
param List contendo esses parametros nomeados
}
\description{
Funcao para criar um objeto com os parametros constantes para rodar o SMAP
}
