% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_SMAPmes.R
\name{SMAPMes.rodadaMensal}
\alias{SMAPMes.rodadaMensal}
\title{Funcao para o calculo do SMAP mensal}
\usage{
SMAPMes.rodadaMensal(param, inic, ETo, P, mRbind, dataM, mComp)
}
\arguments{
\item{param}{Objeto param obtido pela funcao SMAP.param}

\item{inic}{Objeto inic obtido pela funcao SMAP.inic}

\item{ETo}{Valor de ETo do mes}

\item{P}{Valor de precipitacao do mes}

\item{mRbind}{Matriz com iteracoes passadas, onde os resultados do mes serao adicionados,
contendo os valores de Rsolo e Rsub da itera??o anterio}

\item{dataM}{[Opcional] Nomea a linha da matriz de saida com a data fornecida}

\item{mComp}{[Opcional] Altera a matriz de saida, adicionando os resultados das variaveis utilizadas
no calculo}
}
\value{
Matriz com a vaz?o calculada e par?metros Rsolo e Rsub necess?rios para proxima iteracao
}
\description{
Funcao para o calculo do SMAP mensal
}
