% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_SMAPmes.R
\name{SMAPMes.serieMensalQ}
\alias{SMAPMes.serieMensalQ}
\title{Funcao para calcular uma serie de vazoes mensais (m3s)}
\usage{
SMAPMes.serieMensalQ(param, inic, SerieETo, SerieP, dataM)
}
\arguments{
\item{param}{Objeto do tipo SMAP.param}

\item{inic}{Objeto do tipo SMAP.inic}

\item{SerieETo}{Serie mensal de ETo (mm)}

\item{SerieP}{Serie Mensal de P (mm)}

\item{dataM}{Serie com as datas das observacoes}
}
\value{
Vetor com as series de vaz?es
}
\description{
Funcao para calcular uma serie de vazoes mensais (m3s)
}
