% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_SMAPmes.R
\name{SMAPMes.serieMensalQvet}
\alias{SMAPMes.serieMensalQvet}
\title{Serie de vazoes a partir de entrada vetorial}
\usage{
SMAPMes.serieMensalQvet(v, inic, SerieETo, SerieP, dataM)
}
\arguments{
\item{v}{vetor com os parametros do SMAP}

\item{inic}{Objeto do tipo SMAP.inic}

\item{SerieETo}{Serie mensal de ETo (mm)}

\item{SerieP}{Serie Mensal de P (mm)}

\item{dataM}{Serie com as datas das observacoes}
}
\value{
Vetor com as series de vazoes
}
\description{
Serie de vazoes a partir de entrada vetorial
}
