% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_auxiliares.R
\name{Verif_CAVs_ordcres}
\alias{Verif_CAVs_ordcres}
\title{Verificar Ordem Crescente em Colunas de um Data Frame}
\usage{
Verif_CAVs_ordcres(CAVs)
}
\arguments{
\item{CAVs}{Data frame contendo colunas a serem verificadas.}
}
\value{
Vetor de IDs que apresentam colunas fora de ordem.
}
\description{
Identifica IDs que possuem colunas não ordenadas crescentemente.
}
